Extract module specs to shared examples
