+++ spec/spec_helper.rb
@@ -12,7 +12,11 @@ require "pry"
 require "rack"
 require "timecop"
 
+require "support/shared_contexts/token_class"
+require "support/shared_examples/builder"
+require "support/shared_examples/configurable"
 require "support/shared_examples/default_claims"
+require "support/shared_examples/verifier"
 require "support/timecop_helper"
 
 RSpec.configure do |config|
